WIP - file router / vite plugin refactoring
